requirements/base.txt
django>=5.0,<6.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
django-environ>=0.11.0
redis>=5.0.0
celery>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-ratelimit>=4.1.0
drf-spectacular>=0.26.0
python-json-logger>=2.0.0
sentry-sdk>=1.30.0
django-extensions>=3.2.0
factory-boy>=3.3.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
pip-audit>=2.6.0
django-htmx>=1.17.0
reportlab>=4.0.0
openpyxl>=3.1.0
django-celery-beat>=2.5.0
